control-plane/genai-engine/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
openai==1.35.0
pydantic==2.5.0
numpy==1.24.3
pandas==2.1.3
requests==2.31.0
python-json-logger==2.0.7
python-dotenv==1.0.0
httpx==0.25.0

